requirements.txt
aiohttp==3.8.4
aiosignal==1.3.1
asgiref==3.6.0
async-timeout==4.0.2
attrs==22.2.0
certifi==2022.12.7
cffi==1.15.1
charset-normalizer==3.0.1
cryptography==40.0.2
deepl==1.14.0
defusedxml==0.7.1
Django==4.1.7
django-material-admin==1.8.6
django-widget-tweaks==1.4.12
frozenlist==1.3.3
idna==3.4
multidict==6.0.4
mysqlclient==2.2.0
oauthlib==3.2.2
openai==0.27.4
Pillow==9.5.0
pycparser==2.21
PyJWT==2.6.0
pypng==0.20220715.0
python-dateutil==2.8.2
python-dotenv==1.0.0
python3-openid==3.2.0
pytz==2022.7.1
pyzbar==0.1.9
qrcode==7.4.2
requests==2.28.2
requests-oauthlib==1.3.1
six==1.16.0
sqlparse==0.4.3
stripe==5.2.0
tqdm==4.65.0
typing_extensions==4.5.0
urllib3==1.26.14
yarl==1.8.2
gunicorn>=20.0.4

